change login api routes (see API.md), refactor login function, change package manager to yarn
