--- backend/db/.env
@@ -1,3 +1,0 @@
-CONNECTION_STRING : 'mongodb+srv://hau165:Minhhau165@cluster0.zdekj4w.mongodb.net/BanKhoaHoc?retryWrites=true&w=majority'
-PORT : 8000
-SECRET_KEY: "super-secret-key"
