+++ backend/server.js
@@ -6,6 +6,7 @@ const cartRoutes = require('./routes/cartRoutes');
 const categoriesRoutes = require('./routes/categoriesRoutes');
 const topicsRoutes = require('./routes/topicsRoutes');
 const usersRoutes = require('./routes/usersRoutes');
+const loginRoutes = require('./routes/loginRoutes');
 const lecturersRoutes = require('./routes/lecturersRoutes');
 const adminRoutes = require('./routes/adminRoutes');
 const topWeeksRoutes = require('./routes/topWeeksRoutes');
@@ -25,6 +26,7 @@ app.use('/api/cart', cartRoutes);
 app.use('/api/categories', categoriesRoutes);
 app.use('/api/topics', topicsRoutes);
 app.use('/api/users', usersRoutes);
+app.use('/api', loginRoutes);
 app.use('/api/lecturers', lecturersRoutes);
 app.use('/api/admin', adminRoutes);
 app.use('/api/topWeeks', topWeeksRoutes);
